requirements.txt
absl-py==0.12.0
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blis==0.7.4
breadability==0.1.20
bs4==0.0.1
bson==0.5.10
cachetools==4.2.2
cargo==0.3
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cymem==2.0.5
Cython==0.29.23
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==2.2.12
django-widget-tweaks==1.4.8
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
example==0.1.0
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
gast==0.4.0
gensim==4.0.1
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.38.0
gTTS==2.2.2
gunicorn==20.1.0
h5py==3.2.1
idna==3.2
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
razorpay==1.2.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
smart-open==5.1.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.15
sqlparse==0.3.1
srsly==2.4.1
sumy==0.8.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.1
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
thinc==8.0.3
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.1
wrapt==1.12.1
